+++ plugin/src/main/kotlin/org/gradle/github/dependencygraph/internal/util/GradleExtensions.kt
@@ -6,7 +6,7 @@ import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.ProviderFactory
 import org.gradle.internal.operations.BuildOperationListenerManager
 
-internal abstract class PluginCompanionUtils {
+internal abstract class GradleExtensions {
 
     inline val Gradle.objectFactory: ObjectFactory
         get() = service()